+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - Metadata 1.2 support
+- Fix admin query logging [#0006999](https://bugs.oxid-esales.com/view.php?id=6999)
+
  
 ## [6.4.0] - 2019-08-02